0721 랜드마크 등록, 진료 기록 등록

+++ src/main/java/model/LandmarkVO.java
@@ -23,15 +23,15 @@ public class LandmarkVO {
 	private int r_mid_y;      
 	private int r_bot_x;         
 	private int r_bot_y;         
-	private int l_femur_len;       
-	private int l_tibia_len;      
-	private int l_total_len;     
-	private int r_femur_len;     
-	private int r_tibia_len;      
-	private int r_total_len;      
-	private int lr_femur_gap;      
-	private int lr_tibia_gap;      
-	private int lr_total_gap;      
+	private Float l_femur_len;       
+	private Float l_tibia_len;      
+	private Float l_total_len;     
+	private Float r_femur_len;     
+	private Float r_tibia_len;      
+	private Float r_total_len;      
+	private Float lr_femur_gap;      
+	private Float lr_tibia_gap;      
+	private Float lr_total_gap;      
 	private String landmark_date;    
 
 }
